requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
pymongo==4.5.0
gunicorn==21.2.0
python-dotenv==1.0.0
blinker==1.6.2
itsdangerous==2.1.2
Flask-Session==0.5.0
pytz==2023.3
dnspython==2.4.2
flask-pymongo==2.3.0
python-dateutil==2.8.2 